backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "celery>=5.6.2",
    "chromadb>=1.5.2",
    "fastapi>=0.134.0",
    "httpx>=0.28.1",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.10",
    "openai>=2.24.0",
    "passlib[bcrypt]>=1.7.4",
    "psutil>=7.2.2",
    "pydantic[email]>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pypdf>=6.7.4",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "redis>=7.2.1",
    "reportlab>=4.4.10",
    "sentry-sdk[fastapi]>=2.53.0",
    "sqlalchemy[asyncio]>=2.0.47",
    "uvicorn[standard]>=0.41.0",
    "langchain-core>=1.2.16",
    "langchain-text-splitters>=1.1.1",
    "python-docx>=1.2.0",
    "anthropic>=0.84.0",
    "langchain-anthropic>=1.3.4",
    "google-genai>=1.56.0",
    "langchain-google-genai>=4.2.1",
    "authlib>=1.6.8",
    "prometheus-client>=0.24.1",
    "opentelemetry-api>=1.39.1",
    "opentelemetry-sdk>=1.39.1",
    "opentelemetry-instrumentation-fastapi>=0.60b1",
    "opentelemetry-instrumentation-sqlalchemy>=0.60b1",
    "opentelemetry-instrumentation-redis>=0.60b1",
    "opentelemetry-exporter-otlp>=1.39.1",
    "psycopg2>=2.9.11",
    "bcrypt==4.3.0",
]

[dependency-groups]
dev = [
    "freezegun>=1.5.5",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "pytest-xdist>=3.8.0",
]
